Created the basic structure

+++ src/main/java/com/Ankit/EmployeManagement/model/Employee.java
@@ -23,7 +23,7 @@ public class Employee {
     @Email
     private String email;
 
-    private BigDecimal salary;
+    private Double salary;
 
     private LocalDate joiningDate;
